changed type casting to ensure text to integer works

+++ data_warehouse_sql/schema/2_valid_zone/functions/write_sf_account_raw_to_valid.sql
@@ -1356,7 +1356,7 @@ BEGIN
         CAST(title_i AS BOOLEAN) AS title_i,
         type,
         lease_holder,
-        CAST(youth_population AS INTEGER) AS youth_population,
+        CAST(youth_population AS NUMERIC) AS youth_population,
         ys_report_chapter_affiliation,
         CAST(is_deleted AS BOOLEAN) AS is_deleted,
         CAST(sf_created_timestamp AS TIMESTAMPTZ) AS sf_created_timestamp,
